Changed to Byte

+++ src/com/company/Card.java
@@ -11,10 +11,10 @@ public class Card {
         this.symbol = symbol;
     }
 
-    public int score() {
+    public byte score() {
         if (Arrays.asList("+2", "Reverse", "Skip").contains(symbol)) return 20;
         else if (color.equals("Wild")) return 40;
-        else return Integer.parseInt(symbol);
+        else return Byte.parseByte(symbol);
     }
 
     @Override
